requirements.txt
jwcrypto
canonicaljson
py-multihash
requests
PyLD
python-jose[cryptography]